Expunge circleci from cni repo (#187)
Removing circleci as it is no longer necessary for CNI testing. If
someone steps up to maintain IPv6 support and IPv6 support cannot
be implemented in PROW, CircleCI may be brought back into the
implementation. Until that happens, this check is just failing and
there is nothing worse than code that doesn't work ;)