Use su-sxec to chown /home within build container (#186)
* Use su-sxec to chown /home within build container

The home directory is owned by root.  When running in a non-root
context, Kind will be unable to write to the /home directory.  It
is necessary to chown the directory to $(UID) but only when
BUILD_WITH_CONTAINER=1 is true.

Depends-On: https://github.com/istio/common-files/pull/91

* Consolidate other reviews

* Latest build-tools image

`make prow-e2e` nearly works outside of prow. There is a defect
I don't have time to rectify at the moment.

After this PR, there are no regressions in make targets.  The
make targets that fail failed with or without buildtools.