Perform preemption check only when used as a chained plugin
When Istio CNI is not used as a chained plugin, throwing an
error or logging the preemption warning doesn't make sense.
For example, when using Multus CNI, multiple CNI config
files are in the same dir, but none of them has precedence
over the others, as they are loaded by name. In that case,
logging the warning isn't appropriate.